+++ src/lib/ds/treeds.h
@@ -739,7 +739,7 @@ struct ipAuxRbTreeNode{
 
 struct itHashmap{
 	hash_map<unsigned int,itNode *> indexNodeMap;
-	hash_map<uint32_t,unsigned int> nodeIndexMap;
+	hash_map<intptr_t,unsigned int> nodeIndexMap;
 };
 
 #endif /* TREEDS_H_ */
