+++ src/lib/utils/dillutil.h
@@ -83,11 +83,11 @@ public:
 			return null;
 		}
 		hash_map<unsigned int,dillNode *> indexNodeMap;
-		hash_map<uint32_t,unsigned int> nodeIndexMap;
+		hash_map<intptr_t,unsigned int> nodeIndexMap;
 		hash_map<unsigned int,dillNode *>::iterator itToIndexNodeMap;
-		hash_map<uint32_t,unsigned int>::iterator itToNodeIndexMap;
+		hash_map<intptr_t,unsigned int>::iterator itToNodeIndexMap;
 		while(ptr != null){
-			nodeIndexMap.insert(pair<uint32_t,unsigned int>((uint32_t)ptr,startIndex));
+			nodeIndexMap.insert(pair<intptr_t,unsigned int>((intptr_t)ptr,startIndex));
 			indexNodeMap.insert(pair<unsigned int,dillNode *>(startIndex,ptr));
 			startIndex += 1;
 			ptr = ptr->next;
